requirements.txt
youtube.py
python-vlc
pafy
youtube_dl

